fix(marketplace-api): remove calculation floor diff (#386)
marketplace-api: remove calculation floor diff

+++ ark-marketplace-api/src/models/token.rs
@@ -8,7 +8,7 @@ pub struct TokenData {
     pub contract: Option<String>,
     pub token_id: Option<String>,
     pub last_price: Option<BigDecimal>,
-    pub floor_difference: Option<BigDecimal>,
+    pub floor_difference: Option<i32>,
     pub listed_at: Option<i64>,
     pub owner: Option<String>,
     pub minted_at: i64,
